Refactor to allow embedding as framework within other applications
The main changes affect floppy/node.py, floppy.nodeLib.py and
floppyUi.py. The _NODECLASSES, NODECLASSES and MANAGEDNODES are
refactoeded to MappingProxyType style read only views of the _nodeclasses
dictionary used by MetaNode to store and manage all node classes.
The MetaNode.makeManaged and MetaNode.updateManaged are the only means
to create, rename safe and update run and setupmethods of managed
nodes. To explicitly store a managed node the nosafe keyword argument of
the MetaNode.makeManaged method has to be set to false, which is the
default.
Any change to an existing managed node class, including calls to
<NodeClass>._addInput, <NodeClass>._addOutput, <NodeClass>._removeInput,
<NodeClass>._removeOutput as well as a newly created
managed node class can be discared by calling
MetaNode.revertUnsavedManaged method unless this call is preceeded by a
call to MetaNode.makeManaged with nosafe false.

Inputs and outputs are now accessible by i_<InputName> and
o_<OutputName> properties. The getters return the corresponding
InputInfo and OutputInfo objects. Calling the value attributes of these
properties (i_<InputName>.value, o_<OutputName>.value) is eqivalent to
calling _<InputName> on inputs. Assiging a value to these
properties (i_<InputName> = <value>, o_<OutputName> = <value>) is
equivalent to calling _<OutputName>(value) on outputs. This results in
symetric usage of inputs and outputs. The unchecked value of an input
can be accessed by i_<InputName>.r_value.

The __inputs__ and __outputs__ magic attributes expose a
MappingProxyType like concises readonly view of all inputs defined by
the <NodeClass> and its parent classes. To access the inputs and outputs
defined on the <NodeClass> are exposed by the __classinputs__ and
__classoutputs__ magic attributes. Any of these should be
modified only by MetaNode, Node class and related methods definde in
node.py only.

The setupNodesPathes method allows to load nodes from other pathes than
floppy/CustomNodes and creates managedNodes.dat within the specified
savepath if not existing implicitly. Exempt Node, ControlNode, ProxyNode
and SubGraph Node all other nodes were moved to
floppy/Nodes/floppyBaseNodes.py file.
The latter may be mereged lateron with CustomNodes path as soon as
seetupNodesPathes is capable of filtering Node definitions according to
individual .py files and not just by directories with <:cwd:>, <:main:> and
<:floppy> beeing replaced by <current working directory>,
<main script file directory> and <pathtofloppyinstalation>/floppy/ pathes.
In the latter three Node definition files are only located within
CustomNodes and for <:main:> and <:floppy:> also within Nodes
subdirecotry. Any other path specified is assumed to contain
NodeDefinitions directly. If savepath is pecified it is checke if it is
read and writeable and contains a managedNodes.dat file which is
implicitly created if not contained.

On FloppyUi NodeWizard only inoputs and outputs can be removed or added
which are directly defined by the edited managed node class. The dummy
Input and Output now are of NoneType which is painted in gray. They are
removed as soon as a real Inputt and or Output has been added to the
Node. And readded if all inputs and outputs exempt the Trigger input
defined by Node class remain.

If the runCode and setupCode strings only contain any of the follwoing
    super(type(self),self).setup()
    super(type(self),self).run()
    super(self.__class__,self).setup()
    super(self.__class__,self).run()
    super(<classname>,self).setup()
    super(<classname>,self).run()
    <baseclassname>.setup(self)
    <baseclassname>.run(self)

    with:
        <classname>     := self.__class__.__name__
        <baseclassname> := self.__class__.__bases__[0].__name__

Than no explicit run or setup method is created for the managed Node.

In addition the follwoing fixes were made:

1)removed calls to super(<class>,self).__getattibute__ as __getattribute__ if it
does not fined the attribute will raise an AttributeError which will
cause pythong to call __getattr__ again resulting in an infinit loop. If
the base class defines a __getattr__ method than it would be better to
call super(<class>,self).__getattr__ from within __getattr__

2) In floppyUi.Painter2D.PaintEvent changed the line

if issubclass(type(node),DynamicNode):
    node.probeGraph()

to

if issubclass(type(node),SubGraph)
    node.probeGraph()

as the method probeGraph() is defined by SubGrap and not its parent
class DynamicNode.

3) In floppyUi confirmName did not call QWidget.deleteLater() which
caused floppy to crash when  returning to editin inputs after renaming
the edited Node and trying to add the spacer and saveButton widgets to
the new mask.